ci: Temporarily disable `semantic-release` to publish latest version to GitHub
Because the last `Typist Package Release` workflow run failed while
publishing to GitHub Packages, we need to comment out the
`semantic-release` step so that we can run the workflow to publish the
latest package version to GitHub Packages only.

This commit should be reverted as soon as the latest release is
successfully published to GitHub Packages.