feat(extensions)!: Overwrite built-in List/Ordered toggle functions with a `smartToggle` option
BREAKING CHANGE: The `smartToggleBulletList` and
`smartToggleOrderedList` commands were renamed to have the same name as
the built-in toggle functions so that they can easily be used by the
default keyboard shortcuts without having to change the
`addKeyboardShortcuts` function.

The `BulletList` and `OrderedList` extensions now take an additional
option, `smartToggle` (default: `false`), that indicates whether hard
breaks should be replaced by paragraphs before toggling the
selection into a bullet/ordered list, or not.