refactor: General improvements for TypeScript types
BREAKING CHANGE: Fix TypeScript types for the Suggestion plugin
`command` function (allowing for generic override).

Although this was not causing any issues for us due to the convoluted
implementation, the same fix was recently applied to Tiptap, and this
aligns our suggestion factory function implementation with the original
one. Unfortunately, this comes as a breaking change (examples were also
updated to reflect the required changes).

Additionally, other smaller TypeScript types were refactored for
consistency and clarity, but these are safe changes.